saiteki-worker/pyproject.toml
[project]
name = "saiteki-worker"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "openai",
    "numpy",
    "pandas",
    "anthropic",
    "requests",
    "boto3",
    "pydantic",
    "backoff",
    "python-dotenv",
    "instructor",
    "python-Levenshtein",
    "radon",
    "unidiff",
    "dill",
    "hydra-core==1.3.2",
    "matplotlib",
    "networkx",
    "seaborn",
    "moviepy",
    "scikit-learn",
    "adjustText",
    "markdown",
    "aiofiles",
    "loguru",
    "nats-py>=2.11.0",
]
